Checked arithmetic in the broad-phase resource accounting (RBR-02)
The RB-05 budget decided from size_t counts that could wrap: one box over
a 2^22 x 2^21 x 2^21 grid counted 0 cell items and passed a budget of 100,
and a count of 2^64 + 2^43 was reported as an exact 2^43. The counts behind
every budget decision are now CheckedCount arithmetic (checked products,
n(n-1)/2 dividing the even factor first, saturating sums with a sticky
overflow flag that survives the parallel_reduce join and the per-build
accumulation), so a count that does not fit is "at least SIZE_MAX" -- it
exceeds every limit, SIZE_MAX included -- and BroadPhaseBudgetExceeded says
so (`exact`, "at least") instead of claiming a wrapped value. Cumulative
statistics saturate (candidate_emissions_overflowed).

Grid geometry is validated before any floating-point-to-int conversion:
HashGrid::resize refuses a non-finite domain or an invalid cell size
(std::invalid_argument) and a grid with more than INT_MAX cells along an
axis or more cells than a hash key can index (the new named
BroadPhaseUnrepresentable, raised with or without a budget), hash() and
HashItem work in the key type so keys no longer wrap above 2^31 cells,
box_cell_range checks a box against the domain before the cast, and
build_vertex_boxes / compute_mesh_aabb refuse non-finite positions, which
Eigen's min/max would otherwise mask into finite-looking boxes.

Tests: [checked_count] -- the count-only regressions (a representable
2^63 - 2^42 grid with two boxes exact, three boxes overflowing; the
review's 2^64 grid refused by name), the primitives on synthetic values
(products, joins in every order, pairs up to the largest representable n,
exact and one-over limits, SIZE_MAX and disabled budgets), thread-count
independence, emission counts on synthetic items, geometry validation,
non-finite input through the public builds, an unrepresentable sweep
through Candidates::build leaving no candidates.

+++ python/src/broad_phase/hash_grid.cpp
@@ -8,7 +8,7 @@ void define_hash_grid(py::module_& m)
 {
     py::class_<HashItem>(m, "HashItem")
         .def(
-            py::init<int, int>(),
+            py::init<long, long>(),
             "Construct a hash item as a (key, value) pair.", "key"_a, "id"_a)
         .def(
             "__lt__", &HashItem::operator<,
